Separated TripPlanner GUI and Class
